market/requirements.txt
aio-pika==8.1.1
aiohttp==3.8.1
aioredis==1.3.1
aiormq==6.4.1
aiosignal==1.2.0
alembic==1.8.1
amqp==5.1.1
anyio==3.5.0
appdirs==1.4.4
apturl==0.5.2
argcomplete==2.0.0
asgiref==3.5.0
async-timeout==4.0.2
attrs==21.4.0
auditor==0.7.0
autobahn==22.5.1
Automat==20.2.0
backports.zoneinfo==0.2.1
bcrypt==3.1.7
bddcli==2.9.1
bddrest==2.5.7
billiard==3.6.4.0
blinker==1.4
Brlapi==0.7.0
cas==0.1.0
celery==5.2.7
certifi==2019.11.28
cffi==1.15.0
channels==3.0.3
channels-redis==3.2.0
chardet==3.0.4
charset-normalizer==2.0.12
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
colorama==0.4.3
common-library==0.23.0
constantly==15.1.0
cryptography==37.0.2
cupshelpers==1.0
daphne==3.0.2
dbus-python==1.2.16
defer==1.0.6
distlib==0.3.4
distro==1.4.0
distro-info===0.23ubuntu1
Django==3.2
django-appconf==1.0.3
django-celery-beat==2.3.0
django-celery-results==2.4.0
django-common-helpers==0.9.2
django-cors-headers==3.7.0
django-cron==0.5.1
django-extensions==3.2.0
django-filter==2.4.0
django-imagekit==4.0.2
django-timezone-field==5.0
djangorestframework==3.12.4
djangorestframework-simplejwt==4.6.0
dnspython==2.2.1
-e git+ssh://git@git.xeba.tech/xeba/dolphin.git@8b53bd2debe4e142807db7784d9b889491152a23#egg=dolphin
duplicity==0.8.12.0
easycli==1.9.1
email-validator==1.1.3
entrypoints==0.3
et-xmlfile==1.1.0
Faker==8.1.1
fastapi==0.75.1
fasteners==0.14.1
filelock==3.6.0
frozenlist==1.3.1
future==0.18.2
fuzzywuzzy==0.18.0
gevent==21.12.0
greenlet==1.1.2
gunicorn==20.1.0
h11==0.13.0
hiredis==2.0.0
httplib2==0.14.0
httptools==0.2.0
hyperlink==21.0.0
idna==2.8
importlib-metadata==4.12.0
importlib-resources==5.9.0
incremental==21.3.0
iniconfig==1.1.1
itsdangerous==2.0.1
Jinja2==3.1.1
kavenegar==1.1.2
keyring==18.0.1
kombu==5.2.4
language-selector==0.1
launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.3
lml==0.1.0
lockfile==0.12.2
louis==3.12.0
macaroonbakery==1.3.1
Mako==1.1.0
MarkupSafe==2.1.1
mock==4.0.3
monotonic==1.5
msgpack==1.0.4
multidict==6.0.2
nanohttp==1.13.0
netifaces==0.10.4
numpy==1.23.2
oauthlib==3.1.0
olefile==0.46
openpyxl==3.0.10
orjson==3.6.7
packaging==21.3
pamqp==3.2.0
pandas==1.4.3
paramiko==2.6.0
passlib==1.7.4
pbr==5.8.1
pexpect==4.6.0
pika==1.3.0
pilkit==2.0
Pillow==8.2.0
pkgconfig==1.5.1
platformdirs==2.5.1
pluggy==0.13.1
prompt-toolkit==3.0.29
protobuf==3.6.1
psycopg2-binary==2.9.3
py==1.11.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycairo==1.16.2
pycparser==2.21
pycryptodome==3.15.0
pycups==1.9.73
pydantic==1.9.0
pyexcel==0.6.6
pyexcel-io==0.6.6
pyexcel-xlsx==0.6.0
PyGObject==3.36.0
PyJWT==1.7.1
pymacaroons==0.13.0
pymlconf==2.6.1
PyNaCl==1.3.0
pyOpenSSL==22.0.0
pyparsing==3.0.9
pyRFC3339==1.1
pytest==6.2.3
pytest-django==4.2.0
python-apt==2.0.0+ubuntu0.20.4.7
python-crontab==2.6.0
python-dateutil==2.7.3
python-debian===0.1.36ubuntu1
python-dotenv==0.20.0
python-magic==0.4.27
python-multipart==0.0.5
pytz==2022.2.1
pyxdg==0.26
PyYAML==5.3.1
redis==3.2.0
reportlab==3.5.34
requests==2.27.1
requests-unixsocket==0.2.0
restfulpy==4.6.0
SecretStorage==2.3.1
sentry-sdk==1.0.0
service-identity==21.1.0
simplejson==3.16.0
six==1.14.0
sniffio==1.2.0
SQLAlchemy==1.4.36
sqlalchemy-media==0.17.4
SQLAlchemy-Utils==0.38.3
sqlparse==0.2.4
starlette==0.17.1
stevedore==3.5.0
stop-words==2018.7.23
systemd-python==234
text-unidecode==1.3
texttable==1.6.4
toml==0.10.2
tomli==2.0.1
Twisted==22.4.0
txaio==22.2.1
typing_extensions==4.1.1
tzdata==2022.2
ua-parser==0.15.0
ubuntu-advantage-tools==27.2
ubuntu-drivers-common==0.0.0
ufw==0.36
ujson==4.3.0
unattended-upgrades==0.1
urllib3==1.25.8
usb-creator==0.3.7
user-agents==2.2.0
uvicorn==0.15.0
uvloop==0.16.0
vine==5.0.0
virtualenv==20.13.2
virtualenv-clone==0.5.7
virtualenvwrapper==4.8.4
wadllib==1.3.3
watchgod==0.8.2
wcwidth==0.2.5
websockets==10.2
wsproto==0.15.0
xkit==0.0.0
yarl==1.8.1
zipp==3.8.1
zope.event==4.5.0
zope.interface==5.4.0
